(pure-move) migrate to_dtype
