+++ ttnn/core/tensor/host_tensor_factory.cpp
@@ -10,6 +10,7 @@
 #include <tt-metalium/bfloat8.hpp>
 #include <tt-metalium/distributed_host_buffer.hpp>
 #include <tt-metalium/experimental/tensor/host_tensor.hpp>
+#include <tt-metalium/experimental/tensor/tensor_apis.hpp>
 #include <ttnn/tensor/layout/tensor_layout.hpp>
 #include <ttnn/tensor/tensor_spec.hpp>
 #include <ttnn/tensor/types.hpp>
@@ -74,7 +75,7 @@ HostTensor from_span_impl(std::span<const T> buffer, const TensorSpec& spec, T p
     auto host_buffer = HostBuffer(tensor_impl::encode_tensor_data(tt::stl::make_const_span(buffer), spec, pad_value));
 
     auto res = HostTensor(std::move(host_buffer), buffer_spec, TensorTopology{});
-    return tensor_impl::to_dtype(res, spec.data_type());
+    return tt::tt_metal::to_dtype(res, spec.data_type());
 }
 
 }  // namespace CMAKE_UNIQUE_NAMESPACE
